更新google输入法内核.

+++ libgooglepinyin/include/dictlist.h
@@ -33,7 +33,7 @@ class DictList {
   const SpellingTrie *spl_trie_;
 
   // Number of SingCharItem. The first is blank, because id 0 is invalid.
-  size_t scis_num_;
+  uint32 scis_num_;
   char16 *scis_hz_;
   SpellingId *scis_splid_;
 
@@ -42,9 +42,9 @@ class DictList {
 
   // Starting position of those words whose lengths are i+1, counted in
   // char16
-  size_t start_pos_[kMaxLemmaSize + 1];
+  uint32 start_pos_[kMaxLemmaSize + 1];
 
-  size_t start_id_[kMaxLemmaSize + 1];
+  uint32 start_id_[kMaxLemmaSize + 1];
 
   int (*cmp_func_[kMaxLemmaSize])(const void *, const void *);
